Implement methods of transferring information from an item to a tile entity and back when a enchanted block is broken or placed.

+++ carter-enchantments/src/main/java/net/jingles/enchantments/block/Efficiency.java
@@ -91,7 +91,7 @@ public class Efficiency extends BlockEnchant {
 
     if (!canTrigger(event.getContents().getHolder())) return;
 
-    int chance = (getLevel(event.getContents().getHolder()) * 20) / 100;
+    double chance = (getLevel(event.getContents().getHolder()) * 20) / 100D;
 
     if (Math.random() <= chance) {
       ItemStack item = event.getContents().getIngredient();
